+++ components/Leaderboard/types.ts
@@ -1,0 +1,10 @@
+export interface Player {
+  author: string;
+  username: string;
+  pins: number;
+}
+
+export interface User {
+  author: string;
+  username: string;
+}
