+++ components/Menu/types.ts
@@ -1,0 +1,4 @@
+export interface Props {
+  isOpen: boolean;
+  setOpen: (boolean) => void;
+}
