+++ components/Sidebar/types.ts
@@ -1,0 +1,12 @@
+import type { IPin } from '~/lib/types';
+
+export interface Props {
+  pins: IPin[];
+  isOpen: boolean;
+}
+
+export enum ESortKeys {
+  Latest = 'Latest',
+  Oldest = 'Oldest',
+  Distance = 'Distance'
+}
